Force Update `Icon` component
* Icon component is SSR rendered in the context of the `currentTheme`
from `ThemeContext` which is dark by default and in the SSR stage.

* Client side, when hydration occurs, React does a render and don't see
any changes in the virtual DOM when constructed because this was initially
prerendered in a context (server) React is oblivious to (client).

* Using `key={...}` in combination with `setIconKey` in `React.useEffect`
forces a rerender after components mount, making React to update the DOM
with the correct Icon that was rendered.